+++ cmd/influxd/inspect/export_index/export_index.go
@@ -1,4 +1,4 @@
-package inspect
+package export_index
 
 import (
 	"bufio"