+++ cmd/influxd/inspect/verify_tombstone/verify_tombstone_test.go
@@ -1,4 +1,4 @@
-package inspect
+package verify_tombstone
 
 import (
 	"bytes"